+++ CHANGELOG.md
@@ -20,7 +20,8 @@ CHANGELOG
 - Add `--image-name`, `--config`, `--region` parameters to `build-image` command.
 - Add `delete-image` command with `--name`, `--region`, `--force` parameters.
 - Add `describe-image` command with `--name`, `--region` parameters.
-- Add`list-images` command with `--region`, `--color`parameters.
+- Add `list-images` command with `--region`, `--color`parameters.
+- Add `export-cluster-logs` command.
 - Distribute AWS Batch commands: `awsbhosts`, `awsbkill`, `awsbout`, `awsbqueues`, `awsbstat` and `awsbsub`
   as a separate `aws-parallelcluster-awsbatch-cli` PyPI package.
 - Split head node and compute fleet instance roles and add possibility to configure a different instance role